src/backend/requirements.txt
Django==2.1.7
psycopg2==2.7.6
psycopg2-binary==2.7.7
djangorestframework==3.9.2
django-cors-headers==2.5.2
Pillow>=5.3.0,<5.4.0
pytest-django==3.5.1
flake8>=3.6.0,<3.7.0